Render static answer using websockets
This is a spike to get a basic websockets chat working with a
static answer being sent in chunks over a websocket connection.

Currently it works in the following way:

- A user submits a question
- The backend chunks a static answer into pieces and broadcats it via
the ChatChannel using <conversation_id_question_id> as the stream
- The frontend subscribes the the ChatChannel when it recieves a 201
from the backend on question creation
- The frontend either
  - creates a new message element if it's the first chunk sent since the submission
  - updates the last message elements html with the new chunk
- When the full answer has been sent the backend broadcasts a finished
message and the frontend unsubscribes from the channel

I've also added a very simple stop button to allow a user to cancel
their question while the answer is being sent.