Refactor the batch_process rake task
This provides a number of enhancements to the original rake task added:

- This spawns a number of threads to run the tasks concurrently reducing
  the time spent waiting for results.
- it handles passing task arguments and any necessary escaping of those
  arguments
- it adds the returned data to a predictable JSON object structure `{
  "input": "Question", "output": whatever returned data }`. This allows
  compatibility with tasks like `search_results_for_question` that
  return an array
- it outputs warnings (stderr) emitted by a task (with a filter to stop same
  warnings being repeated)
- it raises errors if tasks return a non-zero status code to indicate a
  sub-task failure.
- it avoids previous risks that questions were minimally escaped and
  thus with certain characters could invoke shell commands.

I did also decide not to do the conditional for INPUT or QUESTION, this
is because I intend in a follow up task to modify the
evaluation:generate_answer task to take an INPUT argument and be
consistent with the others. I also plan to add into that the extra data
we have in `evaluation:generate_report` so that we can then remove
`evaluation:generate_report` and instead expect users to use this task
instead:

```
bundle exec rake evaluation:batch_process[generate_answer,
claude_structured_answer] INPUT_PATH=/path/to/input.yaml
```

---

There's a couple of design decisions made here to call out:

First is the mechanism to call rake tasks, I went with Open3.capture3 as
a means to call a new process and get stdout and stderr. I did
originally want to use Rake::Task[task_name].invoke as this would
probably be faster without the overhead of starting rails for each task.
However I couldn't work out how to run this across threads and have
distinct env vars for the question input. I figured the best way to do
this would be to accept regular rake args too for the things set as env
vars but this seemed like a lot of changes.

Second I used a mutex as a way to fully isolate the point of working
with shared state and yielding a block, so to avoid risks of counts
being wrong or running multiple puts concurrently.

Finally I decided to only show warnings the first time they appear so
anything repetitive doesn't spam the output. This could be a naive
decision, particularly after I killed the mega tedious warning about
prometheus server.